pyproject.toml
[project]
name = "mcp-read-only-sql"
version = "0.1.0"
description = "MCP server for read-only SQL queries supporting PostgreSQL and ClickHouse"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Lukas Lehner", email = "lehner.lukas@gmail.com"}
]
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "clickhouse-connect>=0.7.0",
    "paramiko>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-timeout>=2.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-read-only-sql = "src.server:main"
import-dbeaver = "src.config.dbeaver_import:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-timeout>=2.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]